8th commit

+++ src/components/Cart.jsx
@@ -5,9 +5,11 @@ import axios from 'axios';
 import Button from 'react-bootstrap/Button';
 import Card from 'react-bootstrap/Card';
 import TextField from '@mui/material/TextField'
-
+import {Commet} from 'react-loading-indicators';
+import Swal from 'sweetalert2';
 
 export default function Cart() {
+   const [Loading,setLoading]=useState(false);
   const {userID} = useContext(userContext);
   const [addtocart,setAddtocart] = useState({_id:''});
   const [userCartDetails,setUserCartDetails] = useState([]);
@@ -15,15 +17,40 @@ export default function Cart() {
   useEffect(()=>{
     axios.get(`http://localhost:4000/getCartDetails/${userID}`)
     .then((res)=>setUserCartDetails(res.data.userCart))
-    .catch((err)=>console.error(err));
+    .catch((err)=>console.error(err))
+    .finally(()=>setLoading(true));
   });
 
   const removeFromCart = (id)=>{
+
+      Swal.fire({
+  title: "Are you sure?",
+  text: "You won't be able to revert this!",
+  icon: "warning",
+  showCancelButton: true,
+  confirmButtonColor: "#3085d6",
+  cancelButtonColor: "#d33",
+  confirmButtonText: "Yes, delete it!"
+}).then((result)=>{
+  if(result.isConfirmed){
     const uid = userID;
     console.log(uid,id)
     axios.delete('http://localhost:4000/removeFromCart',{data:{"id":id,"userID":uid}})
-    .then((res)=>console.log(res))
+    .then((res)=>{
+      Swal.fire({
+        position: "top-end",
+      title: "Deleted!",
+      text: "Your item removed from cart",
+      icon: "success",
+      showConfirmButton: false,
+      timer: 1500
+    });
+    })
     .catch((err)=>console.error(err));
+  }
+})
+  
+    
   }
  
 
@@ -34,7 +61,21 @@ export default function Cart() {
   const UpdateDetailsInCart = async (data)=>{
     setAddtocart({_id:''})
     await axios.put("http://localhost:4000/addToCart",{"data":data,"userID":userID})
-    .then((res)=>console.log(res))
+    .then((res)=>{
+      Swal.fire({
+        title: "Details Saved",
+        icon:"success",
+        showClass: {
+          popup: `
+            animate__animated
+            animate__fadeInUp
+            animate__faster
+          `
+        },
+        showConfirmButton: false,
+        timer:1500
+      });
+    })
     .catch((err)=>console.error(err));
   }
 
@@ -45,7 +86,25 @@ export default function Cart() {
       .then((res)=>{
         console.log(res.data[0].count);
 if(data.quantity>res.data[0].count){
-        alert(`the item ${data.name} is exceeded its limit`);
+  
+  Swal.fire({
+  title: `the item ${data.name} is exceeded its limit`,
+  showClass: {
+    popup: `
+      animate__animated
+      animate__fadeInUp
+      animate__faster
+    `
+  },
+  hideClass: {
+    popup: `
+      animate__animated
+      animate__fadeOutDown
+      animate__faster
+    `
+  }
+});
+        
         setNavCount(0);
         
       }else{
@@ -61,6 +120,8 @@ if(data.quantity>res.data[0].count){
       
     })
   }
+
+  if(Loading){
   return (
     <div style={{marginTop:"20px"}}>
      <div style={{ display: "flex", gap: "20px", flexWrap: "wrap","paddingLeft":"15px" }}>
@@ -136,4 +197,14 @@ if(data.quantity>res.data[0].count){
 
     </div>
   )
+    }else{
+    return(
+      <div>
+        <center>
+   <Commet color="#07266e" size="medium" text="Loading..." textColor="#5666c2" />
+        </center>
+       
+      </div>
+    )
+  }
 }
